Tightened up Rails version dependency.
The pessimistic version constraint ~> 2.3.5 actually means

  ['>= 2.3.0', '< 2.4.0']

which is fine for running Rails apps on their own, but not so fine when
we expect a specific Rails version, as we do for serenity-23 and the
Cucumber scenarios in which it is used.  This was all okay until Rails
2.3.8 was released a few days ago.

We should probably update the test applications to Rails 2.3.8 anyway,
but that's a task for a different commit.