Allow serenity-23 to use whatever version of Rails is specified by the Gemfile.
This eliminates unnecessary version constraints in the Rails framework:
it's not necessary to specify which version of Rails you want to use
when you're guaranteed to have only one version available.  It also
helps ensure that integration tests pass against the versions of Rails
that bcsec-rails claims it will work with.