Removed deprecation warning specs: Bcsec::SecuredController is "very obsolete" at this point.
Usage of Bcsec::SecuredController will now be accompanied by a
@RuntimeError@ issued by @fail@.

Interestingly, the standard RSpec pattern of testing for an exception
doesn't work, e.g.

    describe SecuredController do
      describe 'deprecated alias' do
        it 'raises' do
          lambda { ::Bcsec::SecuredController }.should raise_error
        end
      end
    end

That example fails with a "nothing was raised" error.  However, if you
eliminate the expectation and try to resolve @Bcsec::SecuredController@,
you'll get an exception.  Go figure.