Integrate Aker configuration with Rails application configuration. Closes #1.
Summary: This ensures that the application global and per-environment
configurations are loaded, in that order, before the railtie calls
`Aker::Rack.use_in`.

In Rails 3, there are three general segments in the boot process:

* Configure
* Initialize
* Finish

Application global and per-environment configuration loading happens in the
"configure", Application and Railtie initialization happens in "initialize",
and app building (including middleware finalization) and `after_initialize`
blocks happen in "finish" (in that order).

Previously, the recommendation was to to global config in an application
initializer and per-environment config in an `after_initialize` block. This
resulted in the following boot process order for Aker-related items:

* Initialize phase: Plugin initializers run, registering the middleware to use
* Initialize phase: Global configuration loads
* Finish phase: Environment-specific configuration loads

While most of the Aker configuration is not used during boot, the middleware
registration is dependent on the modes selected. The new configuration scheme
(outlined in the README in this commit) results in the following more sane
order:

* Configure phase: Global configuration loads
* Configure phase: Environment-specific configuration loads
* Initialize phase: Plugin initializers run, registering the middleware to use